requirements/base.txt
amqp==5.2.0
anyio==4.4.0
APScheduler==3.6.3
asgiref==3.8.1
async-timeout==4.0.3
autoflake==2.2.0
billiard==4.2.0
black==23.7.0
blinker==1.8.2
cachetools==4.2.2
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
classify-imports==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
distlib==0.3.8
Django==5.0.7
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.19.5
django-recaptcha==3.0.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
filelock==3.15.4
flake8==7.1.0
flower==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.10.0
identify==2.6.0
idna==3.7
inflection==0.5.1
isort==5.12.0
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
nplusone==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.7.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
python-telegram-bot==13.15
pytz==2024.1
pyupgrade==3.10.1
PyYAML==6.0.1
redis==5.0.1
reorder-python-imports==3.10.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tokenize-rt==5.2.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
yesqa==1.5.0
